Use param2-based colorization wherever possible for brick blocks and clay blocks.
Also, this gets rid of the tons and tons of single bricks
and lumps of clay, and gets rid of the related crafts,
replacing them with a simple place-and-paint scheme
for the clay blocks and single-color brick block.

Place a default brick block or default clay block, then
right-click on it with a dye color to paint it.

Digging one of these will return a default brick or clay
block and the dye you painted it with.

For the multi-color brick blocks, craft them using a single
default brick block and three portions of dye.  Digging one
returns a default brick block and three portions of dye.

Multi-color brick blocks are now in the creative inventory.